Update image pullPolicy
As we are stating specific versions for the database using `tags` then the `pullPolicy` should be `IfNotPresent`.

A specified version will cause an error that the image already exists if the image exists locally in the docker images directory for a node. 

`always` is an acceptable policy if not tag is provided and this will always try and fetch `latest` and only pull that down if there is a difference between the registry version and the local copy.